+++ test/deprecated.test.js
@@ -5,10 +5,4 @@ var assert = require('assert'),
 
 describe('deprecated stuff', function() {
 
-  it ('should throw an error when using deprecated function assignment', function () {
-    assert.throws(function () {
-      new math.parse('function f(x) = x^2');
-    }, /Deprecated keyword "function"/);
-  });
-
 });